address selection on refresh fix
